# Add cooperative task yielding