--- a/tokio/src/lib.rs
+++ b/tokio/src/lib.rs
@@ -321,6 +321,8 @@ cfg_process! {
 
 pub mod runtime;
 
+pub(crate) mod coop;
+
 cfg_signal! {
     pub mod signal;
 }