@llogiq 2020-01-23
One could certainly write a proc macro attribute to insert checks in all loops.