@carllerche 2020-01-23
@llogiq well that escalated quickly... 😆