@carllerche 2020-01-23
Also, we discussed this some in [Discord](https://discord.gg/tokio), but IMO the budget should only be decremented when using **leaf** resources and not at branches (like iteration).

The reasoning is that, unlike runtimes such as erlang and golang, we do not have a method by which we can weigh the cost of intermediate points in the task. Also, we are not trying to solve **cpu bound** logic on an asynchronous task, only the case where an asynchronous resource is always ready.