@carllerche 2020-01-23
Finally, I would be interested in hooking this into tokio resources then testing something like [tokio-minihttp](github.com/tokio-rs/tokio-minihttp) w/ `wrk` and see how tail latencies are impacted.