@jonhoo 2020-01-23
@carllerche I'm still not sure whether only tracking leaf futures or also intermediate things is the way to go. While I agree that we (probably) aren't going to preempt CPU-bound tasks this way, I don't see a reason why we explicitly don't want to? It might help alleviate some of the issues with blocking tasks that people currently have to solve with `block_in_place`. Also, the various work that a `Future` implementation has to do in `poll` is CPU work that could conceivably be counter "against" it. I'm honestly not sure.

I'll play around with adding annotations to resources.