@jonhoo 2020-01-23
One awkward thing about annotating leaves _and_ intermediates is that it's usually hard for them to yield _after_ doing work. Take `TcpStream::poll_read` for example. If it _does_ read some bytes, and then the preemption decides that it should yield, it needs somewhere to store the number of bytes read (not an issue in `async` blocks/`fn`s). The alternative is that it checks the preemption counter _before_ doing work, but then you might end up never actually getting around to doing useful work if enough other hits have been counted.