@carllerche 2020-01-23
@jonhoo mostly, at this point, we cannot evaluate the cost of each "branch" future, nor can we guarantee that all branch point swill inject code to update the counter. If one task uses branching logic that updates the counter and another doesn't... that will significantly bias the amount of scheduler time one task gets vs. another...

I would err on the side of "keeping it simpler". Add the counting logic in fewer places (the leaves) and see how things behave.