@carllerche 2020-01-23
> One awkward thing about annotating leaves and intermediates is that it's usually hard for them to yield after doing work

99% of the time, "it doesn't matter" is my guess. One option would be to dec first, and if Pending is hit, inc again, but my guess is that if you hit pending in a branch... even if you decremented the counter once, it probably won't significantly impact behavior.