@carllerche 2020-01-23
@jonhoo well, we are trying to reduce dependencies :) Once we figure out a way to expose it publicly, it can be a feature flag probably so you can limit the `tokio` dependency to *just* the counter.