@jonhoo 2020-01-23
Hmm, so, definitely a performance regression...

---

With `basic_scheduler`:
```console
$ (master) wrk -t10 -c500 -d10 http://0.0.0.0:8080/plaintext
Running 10s test @ http://0.0.0.0:8080/plaintext
  10 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.24ms  502.08us  18.36ms   93.22%
    Req/Sec     9.59k   740.74    29.17k    92.30%
  954162 requests in 10.08s, 117.38MB read
Requests/sec:  94698.33
Transfer/sec:     11.65MB

$ (preempt) wrk -t10 -c500 -d10 http://0.0.0.0:8080/plaintext
Running 10s test @ http://0.0.0.0:8080/plaintext
  10 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.27ms  528.39us  10.62ms   84.95%
    Req/Sec     9.52k     0.85k   30.26k    88.90%
  947569 requests in 10.07s, 116.57MB read
Requests/sec:  94125.25
Transfer/sec:     11.58MB
```
---

With `threaded_scheduler`:

```console
$ (master) wrk -t10 -c500 -d10 http://0.0.0.0:8080/plaintext
Running 10s test @ http://0.0.0.0:8080/plaintext
  10 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   573.78us  410.32us  23.54ms   93.47%
    Req/Sec    78.42k    15.86k  112.77k    52.65%
  7808327 requests in 10.10s, 0.94GB read
Requests/sec: 773112.94
Transfer/sec:     95.11MB

$ (preeempt) wrk -t10 -c500 -d10 http://0.0.0.0:8080/plaintext
Running 10s test @ http://0.0.0.0:8080/plaintext
  10 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   648.79us    0.89ms  38.38ms   99.35%
    Req/Sec    73.99k    16.40k  104.67k    49.65%
  7367111 requests in 10.10s, 0.89GB read
Requests/sec: 729472.40
Transfer/sec:     89.74MB
```

---

On this machine:
```console
$ lscpu
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
Byte Order:                      Little Endian
Address sizes:                   46 bits physical, 48 bits virtual
CPU(s):                          40
On-line CPU(s) list:             0-39
Thread(s) per core:              2
Core(s) per socket:              10
Socket(s):                       2
NUMA node(s):                    2
Vendor ID:                       GenuineIntel
CPU family:                      6
Model:                           63
Model name:                      Intel(R) Xeon(R) CPU E5-2660 v3 @ 2.60GHz
Stepping:                        2
CPU MHz:                         2597.092
CPU max MHz:                     2600.0000
CPU min MHz:                     1200.0000
BogoMIPS:                        5196.60
Virtualization:                  VT-x
L1d cache:                       640 KiB
L1i cache:                       640 KiB
L2 cache:                        5 MiB
L3 cache:                        50 MiB
NUMA node0 CPU(s):               0,2,4,6,8,10,12,14,16,18,20,22,24,26,28,30,32,34,36,38
NUMA node1 CPU(s):               1,3,5,7,9,11,13,15,17,19,21,23,25,27,29,31,33,35,37,39
```